Add a profile-logic/data-structures.js file
