+++ src/test/components/ProfileCallTreeView.test.js
@@ -11,11 +11,11 @@ import { Provider } from 'react-redux';
 import mockCanvasContext from '../fixtures/mocks/canvas-context';
 import { storeWithProfile } from '../fixtures/stores';
 import { getBoundingBox } from '../fixtures/utils';
+import { getProfileFromTextSamples } from '../fixtures/profiles/processed-profile';
 import {
-  getProfileFromTextSamples,
   getEmptyThread,
   getEmptyProfile,
-} from '../fixtures/profiles/processed-profile';
+} from '../../profile-logic/data-structures';
 import {
   changeCallTreeSearchString,
   changeImplementationFilter,
